🔒 chore(AuctionsController.cs): add authorization attribute to CreateAuction, UpdateAuction, and DeleteAuction methods to restrict access to authenticated users only 🐛 fix(AuctionsController.cs): set auction.Seller to User.Identity.Name instead of hardcoded value "test" to associate the auction with the authenticated user 🔒 chore(AuctionsController.cs): add check to UpdateAuction and DeleteAuction methods to ensure that the authenticated user is the seller of the auction before allowing updates or deletions

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -46,11 +47,14 @@ public class AuctionsController : ControllerBase
     return _mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<Auction>> CreateAuction(CreateAuctionDto auctionDto)
   {
     var auction = _mapper.Map<Auction>(auctionDto);
-    auction.Seller = "test";
+
+    auction.Seller = User.Identity.Name;
+
     _context.Auctions.Add(auction);
 
     var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -63,6 +67,7 @@ public class AuctionsController : ControllerBase
 
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
   }
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
   {
@@ -70,7 +75,8 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    // TODO check seller == username
+    if (auction.Seller != User.Identity.Name) return Forbid();
+
     auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
     auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -84,17 +90,24 @@ public class AuctionsController : ControllerBase
 
     return BadRequest(result);
   }
+
+
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
     var auction = await _context.Auctions.FindAsync(id);
+
     if (auction == null) return NotFound();
+
+    if (auction.Seller != User.Identity.Name) return Forbid();
+    
     _context.Auctions.Remove(auction);
 
     await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
     var result = await _context.SaveChangesAsync() > 0;
-    
+
     if (!result) return BadRequest("Couldnt update DB");
     return Ok();
